backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Werkzeug==3.0.1
python-dotenv==1.0.0
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
bcrypt==4.0.1